Map labels/annotations to custom vars w/o prefix
It somethimes is requried to define a variable in a Prometheus alert
which gets copied over to Icinga without any prefix. This commit extends
the logic how custom service variables get added by checking if the
label or annotation key startsi with `icinga_`. If this is the case, it
gets added to the custom service variable without any prefix. This is
in addition to the variable being added with the `label_` or
`annotation_prefix`.

Relates APPU-2245