Change Signalilo to not process "resolved" alerts for heartbeats
Alertmanager sends a resolved alert for a missing "heartbeat" alert,
e.g. when Prometheus is not running.  If we process resolved heartbeats
using the code for regular alerts, we set dummy_state to 0 which
prevents the missing heartbeat from showing up as critical in Icinga.

Also split out function createServiceData from updateOrCreateService.