Support for injecting variables with static value on all Icinga2 services
There may be a use case for having a mechanism which ensures that some
set of variables is present on all Icinga2 services created by
Signalilo, e.g. if the Icinga2 server configuration is written in such a
way to use some variables for internal logic.

This commit introduces a new command line flag / environment variable
`--icinga_static_service_var`/`SIGNALILO_ICINGA_STATIC_SERVICE_VAR`
which can be specified any number of times (use newlines in the
environment variable) to define a set of variables and their static
values that must be present on all services.

The implementation will not overwrite variables which already exist on
the incoming alert with the static value from the command line flag.